+++ tests/functional/environment.py
@@ -9,7 +9,7 @@ from tests.functional.utils.context_utils import (
     initialize_scenario_data,
     patch_context
 )
-from tests.functional.utils.db_utils import (
+from tests.functional.utils.test_api_utils import (
     delete_supplier_data_from_sso,
     delete_supplier_data_from_dir
 )
